Time: 5 ms (66.41%) | Memory: 126.6 MB (82.44%) - LeetSync

+++ 4285-smallest-stable-index-ii/smallest-stable-index-ii.java
@@ -1,10 +1,10 @@
 class Solution {
     public int firstStableIndex(int[] nums, int k) {
         int n = nums.length;
-        long[] maxA = new long[n];
-        long[] minA = new long[n];
-        long max = nums[0];
-        long min = nums[n-1];
+        int[] maxA = new int[n];
+        int[] minA = new int[n];
+        int max = nums[0];
+        int min = nums[n-1];
 
         for(int i=0;i<n;i++){
             max = Math.max(max,nums[i]);
